fighting with background particles
