+++ package.json
@@ -15,6 +15,7 @@
     "@types/react": "^19.0.3",
     "@types/react-dom": "^19.0.2",
     "astro": "^5.1.3",
+    "astro-particles": "^2.10.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.469.0",
@@ -24,6 +25,6 @@
     "tailwind-merge": "^2.6.0",
     "tailwindcss": "^3.4.17",
     "tailwindcss-animate": "^1.0.7",
-    "tsparticles": "^3.7.1"
+    "tsparticles-slim": "^2.12.0"
   }
 }
