+++ src/components/Navigation.astro
@@ -1,8 +1,8 @@
 ---
 const navItems = [
-  { href: '/', text: 'Home' },
-  { href: '/officers', text: 'Officers' },
-  { href: '/faq', text: 'FAQ' },
+  { href: "/", text: "Home" },
+  { href: "/officers", text: "Officers" },
+  { href: "/faq", text: "FAQ" },
 ];
 ---
 
@@ -11,32 +11,38 @@ const navItems = [
     <div class="flex justify-between items-center py-4">
       <a href="/" class="flex items-center space-x-2">
         <div class="flex">
-          <span class="h-8 w-8 bg-[--google-blue] rounded-l-full"></span>
-          <span class="h-8 w-8 bg-[--google-red] rounded-r-full"></span>
+          <span class="h-8 w-8 bg-ggreen rounded-full"></span>
+          <span class="h-8 w-8 bg-gred rounded-full"></span>
+          <span class="h-8 w-8 bg-gblue rounded-full"></span>
+          <span class="h-8 w-8 bg-gyellow rounded-full"></span>
         </div>
         <span class="text-xl font-bold text-gray-800">GDG on Campus</span>
       </a>
       <div class="flex items-center space-x-8">
         <ul class="flex space-x-6">
-          {navItems.map((item) => (
-            <li>
-              <a 
-                href={item.href} 
-                class="text-gray-600 hover:text-[--google-blue] transition-colors"
-              >
-                {item.text}
-              </a>
-            </li>
-          ))}
+          {
+            navItems.map((item) => (
+              <li>
+                <a
+                  href={item.href}
+                  class="text-gray-600 hover:text-[--google-blue] transition-colors"
+                >
+                  {item.text}
+                </a>
+              </li>
+            ))
+          }
         </ul>
-        <a 
-          href="https://discord.gg/m2ZwBCabCq" 
-          target="_blank" 
-          rel="noopener noreferrer" 
+        <a
+          href="https://discord.gg/m2ZwBCabCq"
+          target="_blank"
+          rel="noopener noreferrer"
           class="bg-[#5865F2] hover:bg-[#4752C4] text-white px-4 py-2 rounded-full transition-colors flex items-center space-x-2"
         >
           <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 24 24">
-            <path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515a.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0a12.64 12.64 0 0 0-.617-1.25a.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057a19.9 19.9 0 0 0 5.993 3.03a.078.078 0 0 0 .084-.028a14.09 14.09 0 0 0 1.226-1.994a.076.076 0 0 0-.041-.106a13.107 13.107 0 0 1-1.872-.892a.077.077 0 0 1-.008-.128a10.2 10.2 0 0 0 .372-.292a.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127a12.299 12.299 0 0 1-1.873.892a.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028a19.839 19.839 0 0 0 6.002-3.03a.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419c0-1.333.956-2.419 2.157-2.419c1.21 0 2.176 1.096 2.157 2.42c0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419c0-1.333.955-2.419 2.157-2.419c1.21 0 2.176 1.096 2.157 2.42c0 1.333-.946 2.418-2.157 2.418z"/>
+            <path
+              d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515a.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0a12.64 12.64 0 0 0-.617-1.25a.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057a19.9 19.9 0 0 0 5.993 3.03a.078.078 0 0 0 .084-.028a14.09 14.09 0 0 0 1.226-1.994a.076.076 0 0 0-.041-.106a13.107 13.107 0 0 1-1.872-.892a.077.077 0 0 1-.008-.128a10.2 10.2 0 0 0 .372-.292a.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127a12.299 12.299 0 0 1-1.873.892a.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028a19.839 19.839 0 0 0 6.002-3.03a.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419c0-1.333.956-2.419 2.157-2.419c1.21 0 2.176 1.096 2.157 2.42c0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419c0-1.333.955-2.419 2.157-2.419c1.21 0 2.176 1.096 2.157 2.42c0 1.333-.946 2.418-2.157 2.418z"
+            ></path>
           </svg>
           <span>Join Discord</span>
         </a>
