+++ src/layouts/Layout.astro
@@ -1,7 +1,7 @@
 ---
 import Navigation from "../components/Navigation.astro";
 import Footer from "../components/Footer.astro";
-import ParticlesBackground from "../components/ParticlesBackground.jsx";
+import ParticlesBackground from "@/components/background";
 
 interface Props {
   title: string;
